Fix duplication once and for all (hopefully)
When restored a node, the client set a flag so it wouldn't duplicate it
on that client, but that didn't work for other clients. This approach
introduces a setTimeout which isn't ideal, but it should work solidly.
It's not like a flag was ideal either. I'd say this is less ugly.